Fix java.lang.IllegalArgumentException: Original doesn't exist!
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@16846 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/support/FileLoggerHook.java
@@ -298,8 +298,8 @@ public class FileLoggerHook extends LoggerHook {
 									System.err.println(
 											"Closing alt on change caught " + e);
 								}
-								if(previousFile != null) {
-                                                                        FileUtil.renameTo(latestFile, previousFile);
+								if(previousFile != null && previousFile.exists()) {
+									FileUtil.renameTo(latestFile, previousFile);
 									latestFile.delete();
 								} else {
 									latestFile.delete();
